--- api-reference/api-reference/relay/relay--retrieve.mdx
@@ -1,3 +1,0 @@
----
-openapi: openapi_spec get /relay/{relay_id}
----
