+++ crates/common_utils/src/id_type/relay.rs
@@ -1,5 +1,3 @@
-use std::str::FromStr;
-
 crate::id_type!(
     RelayId,
     "A type for relay_id that can be used for relay ids"
@@ -13,12 +11,3 @@ crate::impl_queryable_id_type!(RelayId);
 crate::impl_to_sql_from_sql_id_type!(RelayId);
 
 crate::impl_debug_id_type!(RelayId);
-
-impl FromStr for RelayId {
-    type Err = error_stack::Report<crate::errors::ValidationError>;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let cow_string = std::borrow::Cow::Owned(s.to_string());
-        Self::try_from(cow_string)
-    }
-}
