+++ crates/router/src/core/relay.rs
@@ -1,5 +1,3 @@
-use std::str::FromStr;
-
 use api_models::relay as relay_models;
 use common_utils::{
     self,
@@ -7,7 +5,6 @@ use common_utils::{
     id_type::{self, GenerateId},
 };
 use error_stack::ResultExt;
-use hyperswitch_domain_models::types;
 
 use super::errors::{self, ConnectorErrorExt, RouterResponse, RouterResult, StorageErrorExt};
 use crate::{
@@ -21,9 +18,7 @@ use crate::{
     },
 };
 
-const IRRELEVANT_PAYMENT_INTENT_ID: &str = "irrelevant_payment_intent_id";
-
-const IRRELEVANT_PAYMENT_ATTEMPT_ID: &str = "irrelevant_payment_attempt_id";
+pub mod utils;
 
 pub async fn relay(
     state: SessionState,
@@ -111,17 +106,15 @@ pub async fn relay_refund(
         hyperswitch_domain_models::router_response_types::RefundsResponseData,
     > = connector_data.connector.get_connector_integration();
 
-    let relay_id = id_type::RelayId::generate();
-
-    let relay_domain =
-        get_relay_domain_model(req, merchant_account.get_id(), profile.get_id(), &relay_id);
+    let relay_domain = get_relay_domain_model(req, merchant_account.get_id(), profile.get_id());
 
     let relay_record = db
         .insert_relay(key_manager_state, &key_store, relay_domain)
         .await
-        .change_context(errors::ApiErrorResponse::InternalServerError)?;
+        .change_context(errors::ApiErrorResponse::InternalServerError)
+        .attach_printable("Failed to insert a relay record in db")?;
 
-    let router_data = construct_relay_refund_router_data(
+    let router_data = utils::construct_relay_refund_router_data(
         &state,
         &connector_account.connector_name,
         merchant_id,
@@ -162,131 +155,6 @@ pub async fn relay_refund(
     Ok(response)
 }
 
-pub async fn construct_relay_refund_router_data<'a, F>(
-    state: &'a SessionState,
-    connector_name: &str,
-    merchant_id: &id_type::MerchantId,
-    connector_account: &domain::MerchantConnectorAccount,
-    relay_record: &hyperswitch_domain_models::relay::Relay,
-) -> RouterResult<types::RefundsRouterData<F>> {
-    let connector_auth_type = connector_account
-        .get_connector_account_details()
-        .change_context(errors::ApiErrorResponse::InternalServerError)
-        .attach_printable("Failed while parsing value for ConnectorAuthType")?;
-
-    let webhook_url = Some(payments::helpers::create_webhook_url(
-        &state.base_url.clone(),
-        merchant_id,
-        connector_name,
-    ));
-
-    let supported_connector = &state
-        .conf
-        .multiple_api_version_supported_connectors
-        .supported_connectors;
-
-    let connector_enum = api_models::enums::Connector::from_str(connector_name)
-        .change_context(errors::ConnectorError::InvalidConnectorName)
-        .change_context(errors::ApiErrorResponse::InvalidDataValue {
-            field_name: "connector",
-        })
-        .attach_printable_lazy(|| format!("unable to parse connector name {connector_name:?}"))?;
-
-    let connector_api_version = if supported_connector.contains(&connector_enum) {
-        state
-            .store
-            .find_config_by_key(&format!("connector_api_version_{connector_name}"))
-            .await
-            .map(|value| value.config)
-            .ok()
-    } else {
-        None
-    };
-
-    let hyperswitch_domain_models::relay::RelayData::Refund(relay_refund_data) = relay_record
-        .request_data
-        .clone()
-        .get_required_value("refund relay data")
-        .change_context(errors::ApiErrorResponse::InternalServerError)?;
-
-    let relay_id_string = relay_record.id.get_string_repr().to_string();
-
-    let router_data = hyperswitch_domain_models::router_data::RouterData {
-        flow: std::marker::PhantomData,
-        merchant_id: merchant_id.clone(),
-        customer_id: None,
-        connector: connector_name.to_string(),
-        payment_id: IRRELEVANT_PAYMENT_INTENT_ID.to_string(),
-        attempt_id: IRRELEVANT_PAYMENT_ATTEMPT_ID.to_string(),
-        status: common_enums::AttemptStatus::Charged,
-        payment_method: common_enums::PaymentMethod::default(),
-        connector_auth_type,
-        description: None,
-        return_url: None,
-        address: hyperswitch_domain_models::payment_address::PaymentAddress::default(),
-        auth_type: common_enums::AuthenticationType::default(),
-        connector_meta_data: None,
-        connector_wallets_details: None,
-        amount_captured: None,
-        payment_method_status: None,
-        minor_amount_captured: None,
-        request: hyperswitch_domain_models::router_request_types::RefundsData {
-            refund_id: relay_id_string.clone(),
-            connector_transaction_id: relay_record.connector_resource_id.clone(),
-            refund_amount: relay_refund_data.amount.get_amount_as_i64(),
-            minor_refund_amount: relay_refund_data.amount,
-            currency: relay_refund_data.currency,
-            payment_amount: relay_refund_data.amount.get_amount_as_i64(),
-            minor_payment_amount: relay_refund_data.amount,
-            webhook_url,
-            connector_metadata: None,
-            reason: relay_refund_data.reason,
-            connector_refund_id: relay_record.connector_reference_id.clone(),
-            browser_info: None,
-            split_refunds: None,
-            integrity_object: None,
-            refund_status: common_enums::RefundStatus::from(relay_record.status),
-        },
-
-        response: Ok(
-            hyperswitch_domain_models::router_response_types::RefundsResponseData {
-                connector_refund_id: relay_record.connector_resource_id.clone(),
-                refund_status: common_enums::RefundStatus::default(),
-            },
-        ),
-        access_token: None,
-        session_token: None,
-        reference_id: None,
-        payment_method_token: None,
-        connector_customer: None,
-        recurring_mandate_payment_data: None,
-        preprocessing_id: None,
-        connector_request_reference_id: relay_id_string.clone(),
-        #[cfg(feature = "payouts")]
-        payout_method_data: None,
-        #[cfg(feature = "payouts")]
-        quote_id: None,
-        test_mode: connector_account.test_mode,
-        payment_method_balance: None,
-        connector_api_version,
-        connector_http_status_code: None,
-        external_latency: None,
-        apple_pay_flow: None,
-        frm_metadata: None,
-        refund_id: Some(relay_id_string),
-        dispute_id: None,
-        connector_response: None,
-        integrity_check: Ok(()),
-        additional_merchant_data: None,
-        header_payload: None,
-        connector_mandate_request_reference_id: None,
-        authentication_id: None,
-        psd2_sca_exemption_type: None,
-    };
-
-    Ok(router_data)
-}
-
 // validate relay request
 pub fn validate_relay_refund_request(
     relay_request: &relay_models::RelayRequest,
@@ -318,8 +186,8 @@ pub fn get_relay_domain_model(
     relay_request: &relay_models::RelayRequest,
     merchant_id: &id_type::MerchantId,
     profile_id: &id_type::ProfileId,
-    relay_id: &id_type::RelayId,
 ) -> hyperswitch_domain_models::relay::Relay {
+    let relay_id = id_type::RelayId::generate();
     hyperswitch_domain_models::relay::Relay {
         id: relay_id.clone(),
         connector_resource_id: relay_request.connector_resource_id.clone(),
@@ -342,7 +210,7 @@ impl ForeignFrom<relay_models::RelayData> for hyperswitch_domain_models::relay::
     fn foreign_from(relay: relay_models::RelayData) -> Self {
         match relay {
             relay_models::RelayData::Refund(relay_refund_request) => {
-                Self::Refund(hyperswitch_domain_models::relay::RelayRefundRequest {
+                Self::Refund(hyperswitch_domain_models::relay::RelayRefundData {
                     amount: relay_refund_request.amount,
                     currency: relay_refund_request.currency,
                     reason: relay_refund_request.reason,
