+++ crates/router/src/routes/app.rs
@@ -54,8 +54,7 @@ use super::verification::{apple_pay_merchant_registration, retrieve_apple_pay_ve
 #[cfg(feature = "oltp")]
 use super::webhooks::*;
 use super::{
-    admin, api_keys, cache::*, connector_onboarding, disputes, files, gsm, health::*, profiles,
-    user, user_role,
+    admin, api_keys, cache::*, connector_onboarding, disputes, files, gsm, health::*, profiles, relay, user, user_role
 };
 #[cfg(feature = "v1")]
 use super::{apple_pay_certificates_migration, blocklist, payment_link, webhook_events};
@@ -592,7 +591,7 @@ impl Relay {
     pub fn server(state: AppState) -> Scope {
         web::scope("/relay")
             .app_data(web::Data::new(state))
-            .service(web::resource("").route(web::post().to(super::relay::relay)))
+            .service(web::resource("").route(web::post().to(relay::relay)))
     }
 }
