+++ crates/router/src/routes/lock_utils.rs
@@ -37,6 +37,7 @@ pub enum ApiIdentifier {
     Recon,
     Poll,
     ApplePayCertificatesMigration,
+    Relay,
 }
 
 impl From<Flow> for ApiIdentifier {
@@ -164,7 +165,8 @@ impl From<Flow> for ApiIdentifier {
             | Flow::RefundsFilters
             | Flow::RefundsAggregate
             | Flow::RefundsManualUpdate
-            | Flow::Relay => Self::Refunds,
+             => Self::Refunds,
+             Flow::Relay => Self::Relay,
 
             Flow::FrmFulfillment
             | Flow::IncomingWebhookReceive
